requirements.txt
attrs
aiohttp
argparse
colorama
dnspython
python-whois-extended
requests
python-nmap
